Solving day 7 challenge More info: https://adventofcode.com/2020/day/7
Recursion at its finest. Not to happy with the datastructure I came up with.
But it got the job done.

Haven't gone to deep on the "structure" of the graph to argue if DFS or BFS
is the better approach here.
Hence I opted in to a "DFS" like approach just for fun and to jog my memory.

Hence I also decided to solve both challenges with one implementation. Not
safe for production!