Solving day 9 challenge More info: https://adventofcode.com/2020/day/9
Quick solution as this puzzle isn't too interesting. Solution for the
first part is based on simple mathematics: We are looking for sum of two
preambles equals to the current element. Hence the current element minus
an element in the preamble equals second integer we need in the preamble.

Part two is a simple loop through the list and popleft if the sum is bigger
than the value we are looking for.