Solving day 10 challenge More info: https://adventofcode.com/2020/day/10
First part wasn't too tricky. Once the adapters are sorted you generate
a list of all diffs between the adapters and then count up the diffs that
are either 1 or 3. Counter implementation in pythong comes in handy here.

Part 2 was all about dynamic programming and realizing that in order to
get all options to the n-th adapter you have to sum over n-i adapters and their
option. In this case i = 3. Should be O(n+m).