Solving day 11 challenge More info: https://adventofcode.com/2020/day/10
Played around with memoization on the first part. Solution is based on
iterating through all seats and counting it's neighbours.

Part 2 was more of the same.

There are a couple of optimizations in my head like exit early when the
seat is empy but there is one neighbour. Similar opt for the other case.