Solving day 13 challenge More info: https://adventofcode.com/2020/day/13
Part 2 was actually a hard problem. Took quite a while until I concluded
that this could be solved with the Chinese remainder theorem.

Based on my manual scribbling it appeared that a) all buses are integrals.
b) every bus has a cycle and we need to combine all cycles to align to
a specific pattern such that Bus 1