Solving day 19 challenge More info: https://adventofcode.com/2020/day/19
Holy moly - these type of grammar / statemachine things are hard to wrap my
head around. In the end the datastructure sets us up to represent the
"statemachine". The actual code was a good chance to dive deepr into python
generators.

The idea is a BFS approach through all possible states and return/yield
in the case we hit a dead end and the word was too long or return an empty
word if we hit a terminal state from the statemachine.

https://en.wikipedia.org/wiki/Tree_automaton