Remove unnecessary readlines()
This somehow slipped through the cracks... well that's how it ends
if you mind wanders and you do copy paste. Nothing a simple find replace
can fix tough.

For reference readlines() normally reads the whole file and puts it in a
list of str... which is not necessary as all implementations in my case work
on a line by line basis.